requirements.txt
fastapi
openai
python-dotenv
uvicorn
requests
slowapi
psycopg2-binary
